moved ingress/egress to own file
